[BugFix] 캐릭터 사망 시 Death 애니메이션으로 전환되지 않던 버그 수정

+++ Assets/Scripts/Unit/GameScene/Units/FSMs/Units/Monster/States/MonsterSkillState.cs
@@ -26,7 +26,7 @@ namespace Unit.GameScene.Units.FSMs.Units.Monster.States
             
             FsmController.RegisterOnAttackEventHandler(OnAttack);
             FsmController.SetBool(MonsterBaseStateInfo.StateParameter, true, ChangeToDefaultState);
-            FsmController.SetInteger(_skillInfo.SkillParameter, _skillInfo.SkillValue, null);
+            FsmController.SetFloat(_skillInfo.SkillParameter, _skillInfo.SkillValue, null);
         }
 
         public override void Exit()
